Finalização de perfil aprendiz

+++ Project/frontend/src/context/authContext.tsx
@@ -1,13 +1,14 @@
 import { createContext, useContext, useState } from "react";
+import Swal from "sweetalert2";
 
 interface Usuario {
     token: string
     user: {
         EDV: number
         tipo: string
+        ativo:boolean
     }
 }
-
 
 interface AuthContextType {
     usuario: Usuario | null;
@@ -22,36 +23,37 @@ export function AuthProvider({children}: {children: React.ReactNode}) {
         const salvo = localStorage.getItem("usuario");
         return salvo ? JSON.parse(salvo) : null;
     });
-
     function login(usuario: Usuario){
+        if (!usuario.user.ativo) {
+            Swal.fire({
+                icon: "error",
+                title: "Usuário inativo",
+                text: "Sua conta está desativada.",
+                confirmButtonColor: "#2B83D5",
+                confirmButtonText: "OK"
+            });
+    
+            return;
+        }
         setUsuario(usuario);
+        console.log("User "+ usuario.user)
         localStorage.setItem(
             "usuario",
-            JSON.stringify(usuario.user)
+            JSON.stringify(usuario)
         );
         localStorage.setItem(
             "token",
             JSON.stringify(usuario.token)
         );
+       
     }
-
     function logout(){
         setUsuario(null);
         localStorage.removeItem("usuario");
         localStorage.removeItem("token");
     }
-
-
     return (
-        <AuthContext.Provider 
-            value={{
-                usuario,
-                login,
-                logout
-            }}
-        >
-            {children}
-        </AuthContext.Provider>
+        <AuthContext.Provider value={{usuario, login, logout}}>{children}</AuthContext.Provider>
     );
 }
